+++ Integrations/ThreatConnect/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+Added 5 new commands.
+   - ***tc-get-group***
+   - ***tc-get-group-attributes***
+   - ***tc-get-group-security-labels***
+   - ***tc-get-group-tags***
 
 
 ## [19.8.2] - 2019-08-22